backend/requirements.txt
django==5.2.11
djangorestframework==3.16.1
psycopg2-binary==2.9.11
gunicorn==25.1.0
anthropic==0.97.0
python-dotenv==1.2.2